requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart

discord.py[voice]
PyNaCl

requests
python-dotenv

passlib[bcrypt]

pydantic
pydantic-settings

aiofiles

matplotlib